轨迹点

+++ trace-baidu/src/main/java/club/mrxiao/baidu/config/BaiduTraceConfig.java
@@ -20,11 +20,11 @@ public class BaiduTraceConfig {
     /**
      *  service的ID
      */
-    private Integer serviceId;
+    private long serviceId;
 
     private Map<String,Object> baseRequest;
 
-    public BaiduTraceConfig(String ak,Integer serviceId){
+    public BaiduTraceConfig(String ak,long serviceId){
         this.ak = ak;
         this.serviceId = serviceId;
         this.baseRequest = new HashMap<String, Object>(3);
